+++ code/ui.datagrid/src/api/Cell.TEST.ts
