+++ code/ui.datagrid/src/api/Cell.ts
