+++ code/ui.datagrid/src/api/Grid.ts
@@ -1,6 +1,6 @@
 import { Subject } from 'rxjs';
 import { filter, map, share, takeUntil } from 'rxjs/operators';
-import { t } from '../../common';
+import { t } from '../common';
 import { Cell } from './Cell';
 
 /**
