+++ code/ui.datagrid/src/components/Grid/hook.keyboard.ts
@@ -1,4 +1,4 @@
-import { Grid } from '../api';
+import { Grid } from '../../api';
 
 /**
  * Factory for creating a grid's `beforeKeyDown` handler.
