+++ code/ui.datagrid/src/components/factory/factory.ts
@@ -1,5 +1,5 @@
 import { t } from '../../common';
-import { Grid } from '../api';
+import { Grid } from '../../api';
 
 const defaultFactory: t.GridFactory = req => null;
